Finished conversion to CSS2
